refactor: move HttpHelper to http folder
