+++ src/main/adapters/ExpressRouteAdapter.spec.ts
@@ -2,7 +2,7 @@ import Controller from '../../presentation/protocols/Controller'
 import HttpRequest from '../../presentation/protocols/HttpRequest'
 import HttpResponse from '../../presentation/protocols/HttpResponse'
 import ExpressRouteAdapter from './ExpressRouteAdapter'
-import HttpHelper from '../../presentation/helpers/HttpHelper'
+import HttpHelper from '../../presentation/helpers/http/HttpHelper'
 
 class ControllerStub implements Controller {
   async handle (_: HttpRequest): Promise<HttpResponse> {
