+++ src/presentation/controllers/login/LoginController.spec.ts
@@ -1,6 +1,6 @@
 import Controller from '../../protocols/Controller'
 import LoginController from './LoginController'
-import HttpHelper from '../../helpers/HttpHelper'
+import HttpHelper from '../../helpers/http/HttpHelper'
 import AuthenticateUser from '../../../domain/usecases/AuthenticateUser'
 import AuthenticateCredentials from '../../../domain/usecases/AuthenticateCredentials'
 import HttpRequest from '../../protocols/HttpRequest'
